requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
openai>=1.0.0
aiofiles>=23.2.1
PyYAML>=6.0.1
Jinja2>=3.1.2
python-dotenv>=1.0.0